Implemented base classes Graph and DiGraph.
Also implemented graph functions and all corresponding tests.
Made the code working properly with the compilers ADVANCED OPTIMIZATIONS
mode.

The compiled output is now wrapped in an immediate function to prevent
polluting the global scope. Only jsnx is globally available and exposes
the same properties as the original networkx symbol (what is ported so far).