Removes __iterator__ from graph classes to fix #5.
Since Firefox implements iterators natively and also recognizes them in
'for...in', the property 'Graph.prototype.__iterator__' causes problems in FF.

Removing it was the easiest solution. That means that graph objects cannot be
directly passed to 'goog.iter' methods, but the 'jsnx.helper' methods have been
updated to handle graph objects separately and through that, they work the
same as if they are still iterables.

An iterator can always be created by passing a graph object to
'jsnx.helper.iter'.