Adds a new Jasmine matcher, toEqualAny.
Some graph operations don't have a unique result. Since the order in
which objects are iterated over is implementation dependent, different
browsers can return different results for the same graph.

This methods accepts a list of values to compare a result against and
passes successfully if the result is equal to at least one of them.