Changes visualization of directed graphs.
Instead of drawing two arcs, two straight lines are drawn with half
arrow heads.
The edges are not drawn as being simple lines, but as paths and
because of this they have to be styled differently now: Instead of using
'stroke' to color an edge, you have to use 'fill'.
'stroke-width' is still used to compute the overall width of the edge,
but the stroke width at the SVG element will be 0 (i.e. it cannot be
styled). This is done to ensure that the arrow heads look properly.