Revert dag test cases.
The test cases couldn't actually pass, because we were renaming the
async methods to `gen_*`. Currently the tests only test the sync method.
We should add tests for the async method as well.

I added a test for `delegateToWorker` to make sure it returns a promise
and handles it correctly. This only tests the functionality in Node
though. We have to get the tests working in the browser as well.

I will add additional tests for the transform to ensure that it works as
intended.