Codemod under_score style to camelCase
... because camelCase is more common in the JS community and I don't
want to incur the wrath of the Node community.

I used recast to make the transformation. The script is included in this
diff. I had to make only minor adjustments for variables of the form
`__foo__`.

All unit tests pass.

If desired, I will create an "migration" file which simply adds back the
underscore API when included (in the browser, I don't think it's
necessary for Node).