Use 6to5 instead of esnext
6to5 has better support for ES6 features, a minimal runtime and allows
us to include the runtime manually where we want to. In contrast, esnext added
a tiny runtime to every file where needed.

Since 6to5 uses es6-shim, we can also make use of the ES6 API (such as
Array.from). However, it will affect the environment, since it is a
polyfill (which could be considered as bad, but I think it's OK).