Enable and simplify custom browser builds
This rearranges some files to make it easier to use JSNetworkX with
`browserify`.

It is now possible to require individual modules and still get the
benefits of async code execution through workers.

Example:

```
var completeGraph = require('jsnetworkx/node/generators/classic').completeGraph;
var cluster = require('jsnetworkx/node/algorithms/cluster');
var WorkerSettings = require('jsnetworkx/node/WorkerSettings');
var initializeBrowserWorker = require('jsnetworkx/node/initializeBrowserWorker');

global.genClustering = cluster.genClustering;
global.completeGraph = completeGraph;
WorkerSettings.methodLookupFunction = function(name) {
  return cluster[name];
};
initializeBrowserWorker();
```

This is a custom module which only exposes `completeGraph` and `genClustering`.

`WorkerSettings.methodLookupFunction`

This function is used by the worker (and `delegateToSync`!) to resolve
the method name the worker receives, to the actual function. E.g. the
prebuilt version of JSNetworkX uses `name => jsnx[name]`.

`initializeBrowserWorker`

This function tests whether we are in a worker environment and if yes,
sets up the event handlers and takes care of (de)serializing the data.
This function should be called in the worker script.

Given the example above, the custom bundle could be created with

  JSNETWORKX_BUNDLE='bundle.js' browserify test.js > bundle.js

In this case, `bundle.js` would be the script included in the page, as
well as the worker script (just like `jsnetworkx.js` is). However, one
could also use two separate scripts. The important part is that the
worker script has access to the *synchronous* versions of the methods
that the page script is going to call (in the example above,
`clustering`).

Overall this allows for custom browser builds that are smaller, but
there doesn't seem to be a big size advantage yet. E.g. the script
above only saves 80KB (unminified) compared to the fill library, due to
the fact the jsnetworkx's core is rather large. But we can optimize this
in the future.